add tests for retrieving words
Added tests. Also cleaned up the rules.
