+++ package.json
@@ -20,6 +20,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@jest/globals": "^29.4.3",
     "@types/jest": "^29.4.0",
     "@types/node": "^18.14.2",
     "@typescript-eslint/eslint-plugin": "^5.53.0",
@@ -29,6 +30,7 @@
     "nodemon": "^2.0.20",
     "prettier": "^2.8.4",
     "rimraf": "^4.1.2",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
   }
