+++ src/data/frenchWordsSource.json
@@ -788,7 +788,7 @@
     "french": "garçon",
     "gender": 0,
     "genderRuleKey": 80,
-    "exception": true
+    "exception": false
   },
   {
     "english": "crime",
@@ -2199,10 +2199,10 @@
   },
   {
     "english": "career",
-    "french": "la",
-    "gender": 0,
-    "genderRuleKey": 72,
-    "exception": true
+    "french": "carrière",
+    "gender": 1,
+    "genderRuleKey": 28,
+    "exception": false
   },
   {
     "english": "marriage",
@@ -3109,9 +3109,9 @@
   },
   {
     "english": "material",
-    "french": "matériau",
+    "french": "matérial",
     "gender": 0,
-    "genderRuleKey": 71,
+    "genderRuleKey": 73,
     "exception": true
   },
   {
@@ -5678,9 +5678,9 @@
   },
   {
     "english": "reading",
-    "french": "la",
+    "french": "reading",
     "gender": 0,
-    "genderRuleKey": 72,
+    "genderRuleKey": 92,
     "exception": true
   },
   {
@@ -6464,7 +6464,7 @@
     "english": "green",
     "french": "vert",
     "gender": 0,
-    "genderRuleKey": 73,
+    "genderRuleKey": 94,
     "exception": true
   },
   {
@@ -9621,8 +9621,8 @@
     "english": "blue",
     "french": "bleu",
     "gender": 0,
-    "genderRuleKey": 71,
-    "exception": true
+    "genderRuleKey": 94,
+    "exception": false
   },
   {
     "english": "engagement",
@@ -12015,7 +12015,7 @@
     "english": "red",
     "french": "rouge",
     "gender": 0,
-    "genderRuleKey": 67,
+    "genderRuleKey": 94,
     "exception": false
   },
   {
@@ -12222,11 +12222,18 @@
     "exception": true
   },
   {
-    "english": "hundred",
+    "english": "about one hundred",
     "french": "centaine",
     "gender": 1,
-    "genderRuleKey": 37,
-    "exception": true
+    "genderRuleKey": 97,
+    "exception": false
+  },
+  {
+    "english": "one hundred",
+    "french": "cent",
+    "gender": 1,
+    "genderRuleKey": 96,
+    "exception": false
   },
   {
     "english": "councillor",
@@ -12789,7 +12796,7 @@
     "exception": true
   },
   {
-    "english": "palm",
+    "english": "palm (of the hand)",
     "french": "paume",
     "gender": 1,
     "genderRuleKey": 54,
@@ -13680,13 +13687,6 @@
   {
     "english": "beak",
     "french": "bec",
-    "gender": 0,
-    "genderRuleKey": 73,
-    "exception": true
-  },
-  {
-    "english": "lead",
-    "french": "plomb",
     "gender": 0,
     "genderRuleKey": 73,
     "exception": true
@@ -15131,6 +15131,125 @@
     "french": "dentifrice",
     "gender": 0,
     "genderRuleKey": 11,
+    "exception": false
+  },
+  {
+    "english": "Bordeaux",
+    "french": "Bordeaux",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Bourgorgne",
+    "french": "Bourgorgne",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Chablis",
+    "french": "Chablis",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Pinot",
+    "french": "Pinot",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Brie",
+    "french": "Brie",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Cantal",
+    "french": "Cantal",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Camembert",
+    "french": "Camembert",
+    "gender": 0,
+    "genderRuleKey": 93,
+    "exception": false
+  },
+  {
+    "english": "Tomme",
+    "french": "Tomme",
+    "gender": 1,
+    "genderRuleKey": 93,
+    "exception": true
+  },
+  {
+    "english": "Brique",
+    "french": "Brique",
+    "gender": 1,
+    "genderRuleKey": 93,
+    "exception": true
+  },
+  {
+    "english": "yellow",
+    "french": "jaune",
+    "gender": 0,
+    "genderRuleKey": 94,
+    "exception": false
+  },
+  {
+    "english": "gold - the metal",
+    "french": "or",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "silver - the metal",
+    "french": "argent",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "lead - the metal",
+    "french": "plomb",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "Chinese - the language",
+    "french": "chinois",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "English - the language",
+    "french": "anglais",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "French - the language",
+    "french": "français",
+    "gender": 0,
+    "genderRuleKey": 98,
+    "exception": false
+  },
+  {
+    "english": "Spanish - the language",
+    "french": "espagnol",
+    "gender": 0,
+    "genderRuleKey": 98,
     "exception": false
   }
 ]
