--- src/example.test.ts
@@ -1,5 +1,0 @@
-describe('My Test Suite', () => {
-  it('My Test Case', () => {
-    expect(true).toEqual(true);
-  });
-});
