+++ src/index.ts
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
-
 const args = process.argv.slice(2);
 console.log('Hitting from another thread works.', args);
+
