--- src/models/genderPatterns.ts
@@ -1,120 +1,0 @@
-// https://frenchtogether.com/french-nouns-gender/
-
-export const endingsPatterns = [
-  { key: 0, gender: 0, rule: 'phone', exceptions: [], explanation: '' },
-  { key: 1, gender: 1, rule: 'ence', exceptions: ['silence'], explanation: '' },
-  { key: 2, gender: 1, rule: 'euse', exceptions: [], explanation: '' },
-  { key: 3, gender: 1, rule: 'ance', exceptions: [], explanation: '' },
-  { key: 4, gender: 1, rule: 'esse', exceptions: [], explanation: '' },
-  { key: 5, gender: 1, rule: 'ique', exceptions: ['graphique', 'périphérique', 'plastique', 'moustique'], explanation: '' },
-  { key: 6, gender: 1, rule: 'ière', exceptions: ['arrière', 'cimetière', 'derrière'], explanation: '' },
-  { key: 7, gender: 1, rule: 'erie', exceptions: [], explanation: '' },
-  { key: 8, gender: 0, rule: 'ueil', exceptions: [], explanation: '' },
-  { key: 9, gender: 0, rule: 'isme', exceptions: [], explanation: '' },
-
-  { key: 10, gender: 0, rule: 'ment', exceptions: ['jument'], explanation: '' },
-  { key: 11, gender: 1, rule: 'rice', exceptions: ['dentifrice'], explanation: '' },
-  { key: 12, gender: 0, rule: 'cide', exceptions: [], explanation: '' },
-
-  { key: 9, gender: 0, rule: /([^s|c][a,e,i,o]){1}n[d|t]+\W/, exceptions: [], explanation: 'Words that end in a pattern of vowel + n + d or t are almost all masculine, unless the ending follows an s or c.' },
-  { key: 24, gender: 0, rule: 'ail', exceptions: [], explanation: '' },
-  { key: 34, gender: 1, rule: 'ace', exceptions: ['ace', 'palace'], explanation: '' },
-  { key: 35, gender: 1, rule: 'ade', exceptions: ['grade', 'jade', 'stade'], explanation: '' },
-  { key: 16, gender: 0, rule: 'age', exceptions: ['cage', 'image', 'nage', 'page', 'plage', 'rage'], explanation: '' },
-  { key: 16, gender: 1, rule: 'aie', exercice: [], explanation: '' },
-  { key: 36, gender: 1, rule: 'ale', exceptions: ['châle', 'pétale', 'scandale'], explanation: '' },
-  { key: 15, gender: 0, rule: 'ard', exceptions: [], explanation: '' },
-  { key: 17, gender: 0, rule: 'ble', exceptions: ['cible', 'étable', 'fable', 'table'], explanation: '' },
-  { key: 18, gender: 0, rule: 'cle', exceptions: ['boucle'], explanation: '' },
-  { key: 26, gender: 0, rule: 'eau', exceptions: ['eau', 'peau'], explanation: '' },
-  { key: 14, gender: 0, rule: 'ège', exceptions: [], explanation: '' },
-  { key: 25, gender: 0, rule: 'eil', exceptions: [], explanation: '' },
-  { key: 27, gender: 0, rule: 'ème', exceptions: [], explanation: '' },
-  { key: 28, gender: 1, rule: 'ère', exceptions: ['père', 'frère'], explanation: '' },
-  { key: 19, gender: 0, rule: 'ing', exceptions: [], explanation: '' },
-  { key: 37, gender: 1, rule: 'ine', exceptions: ['capitaine', 'domaine', 'moine', 'magazine', 'patrimoine'], explanation: '' },
-  { key: 38, gender: 1, rule: 'ion', exceptions: ['avion', 'bastion', 'billion', 'camion', 'cation', 'dominion', 'espion', 'ion', 'lampion', 'lion', 'million', 'morpion', 'pion', 'scion', 'scorpion', 'trillion'], explanation: '' },
-  { key: 39, gender: 1, rule: 'ire', exceptions: ['auditoire', 'commentaire', 'dictionnaire', 'directoire', 'horaire', 'itinéraire', 'ivoire', 'laboratoire', 'navire', 'pourboire', 'purgatoire', 'répertoire', 'salaire', 'sommaire', 'sourire', 'territoire', 'vocabulaire'], explanation: '' },
-  { key: 40, gender: 1, rule: 'ise', exceptions: [], explanation: '' },
-  { key: 41, gender: 1, rule: 'ite', exceptions: ['anthracite', 'ermite', 'granite', 'graphite', 'mérite', 'opposite', 'plébiscite', 'rite', 'satellite', 'site', 'termite'], explanation: '' },
-  { key: 42, gender: 1, rule: 'lle', exceptions: ['braille', 'gorille', 'intervalle', 'mille', 'portefeuille', 'vaudeville', 'vermicelle', 'violoncelle'], explanation: '' },
-  { key: 43, gender: 1, rule: 'mme', exceptions: ['dilemme', 'gramme', 'programme'], explanation: '' },
-  { key: 44, gender: 1, rule: 'nde', exceptions: ['monde'], explanation: '' },
-  { key: 45, gender: 1, rule: 'nne', exceptions: [], explanation: '' }, { key: 21, gender: 0, rule: 'one', exceptions: [], explanation: '' },
-  { key: 23, gender: 0, rule: 'oir', exceptions: [], explanation: '' },
-  { key: 46, gender: 1, rule: 'ole', exceptions: ['contrôle', 'monopole', 'rôle', 'symbole'], explanation: '' },
-  { key: 22, gender: 0, rule: 'ope', exceptions: [], explanation: '' },
-  { key: 47, gender: 1, rule: 'rre', exceptions: ['beurre', 'parterre', 'tonnerre', 'verre'], explanation: '' },
-  { key: 29, gender: 1, rule: 'son', exceptions: ['blouson'], explanation: '' },
-  { key: 20, gender: 0, rule: 'ste', exceptions: ['liste', 'modiste', 'piste'], explanation: '' },
-  { key: 13, gender: 0, rule: 'tre', exceptions: ['fenêtre', 'huître', 'lettre', 'montre', 'rencontre', 'vitre'], explanation: '' },
-  { key: 30, gender: 1, rule: 'tié', exceptions: [], explanation: '' },
-  { key: 31, gender: 1, rule: 'ude', exceptions: ['coude', 'interlude', 'prélude'], explanation: '' },
-  { key: 32, gender: 1, rule: 'ule', exceptions: ['préambule', 'scrupule', 'tentacule', 'testicule', 'véhicule', 'ventricule', 'vestibule'], explanation: '' },
-  { key: 33, gender: 1, rule: 'ure', exceptions: ['centaure', 'cyanure', 'dinosaure', 'murmure'], explanation: '' },
-
-
-  { key: 9, gender: 0, rule: /([a,i,o]){1}n+\W/, exceptions: [], explanation: 'Words that end in a pattern of a, i or o + n are almost all masculine.' },
-  { key: 52, gender: 0, rule: 'at', exceptions: [], explanation: '' },
-  { key: 56, gender: 1, rule: 'be', exceptions: ['cube', 'globe', 'microbe', 'tube', 'verbe'], explanation: '' },
-  { key: 57, gender: 1, rule: 'ce', exceptions: ['artifice', 'armistice', 'appendice', 'bénéfice', 'caprice', 'commerce', 'dentifrice', 'divorce', 'exercice', 'office', 'orifice', 'précipice', 'prince', 'sacrifice', 'service', 'silence', 'solstice', 'supplice', 'vice'], explanation: '' },
-  { key: 58, gender: 1, rule: 'cé', exceptions: ['crustacé'], explanation: '' }, { key: 48, gender: 0, rule: 'de', exceptions: ['bride', 'merde', 'méthode', 'pinède', 'ade, -nde, -ude endings'], explanation: '' },
-  { key: 59, gender: 1, rule: 'ee', exceptions: ['pedigree'], explanation: '' },
-  { key: 49, gender: 1, rule: 'ée', exceptions: ['lycée', 'musée', 'apogée', 'périgée', 'trophée'], explanation: '' },
-  { key: 50, gender: 0, rule: 'et ', exceptions: [], explanation: '' },
-  { key: 61, gender: 1, rule: 'fe', exceptions: ['golfe'], explanation: '' },
-  { key: 62, gender: 1, rule: 'ie', exceptions: ['incendie', 'foie', 'génie', 'parapluie', 'sosie'], explanation: '' },
-  { key: 53, gender: 0, rule: 'il', exceptions: [], explanation: '' },
-  { key: 54, gender: 0, rule: 'me', exceptions: ['alarme', 'âme', 'arme', 'cime', 'coutume', 'crème', 'écume', 'énigme', 'estime', 'ferme', 'firme', 'forme', 'larme', 'plume', 'rame', 'rime'], explanation: '' },
-  { key: 51, gender: 0, rule: 'ot', exceptions: [], explanation: '' },
-  { key: 55, gender: 0, rule: 'ou', exceptions: [], explanation: '' },
-  { key: 63, gender: 1, rule: 'se', exceptions: ['carosse', 'colosse', 'gypse', 'inverse', 'malaise', 'pamplemousse', 'parebrise', 'suspense'], explanation: '' },
-  { key: 64, gender: 1, rule: 'sé', exceptions: ['exposé', 'opposé'], explanation: '' },
-  { key: 65, gender: 1, rule: 'té', exceptions: ['arrêté', 'comité', 'comté', 'côté', 'député', 'été', 'pâté', 'traité'], explanation: '' },
-  { key: 66, gender: 1, rule: 'ue', exceptions: ['abaque'], explanation: '' },
-
-  { key: 72, gender: 0, rule: 'a', exceptions: ['cafétéria', 'pizza'], explanation: '' },
-  { key: 67, gender: 1, rule: 'e', exceptions: ['peuple'], explanation: '' },
-  { key: 69, gender: 0, rule: 'i', exceptions: ['foi', 'fourmi', 'loi', 'paroi', 'merci'], explanation: '' },
-  { key: 70, gender: 0, rule: 'é', exceptions: ['clé', 'psyché', 'sé, té, and tié endings'], explanation: '' },
-  { key: 68, gender: 0, rule: 'o', exceptions: ['dactylo', 'dynamo', 'libido', 'radio'], explanation: '' }, //most of these are apocopes of longer feminine words
-  { key: 71, gender: 0, rule: 'u', exceptions: ['tribu', 'vertu'], explanation: '' },
-];
-
-export const consonantPattern = [
-  { key: 73, gender: 0, rule: 'Words ending in consonants (in the spelling) are usually masculine.', exceptions: ['mob', 'pub', 'alloc', 'collec', 'doc', 'embroc', 'fac', 'réduc', 'soif', 'clef', 'nef', 'façon', 'fin', 'leçon', 'main', 'maman', 'rançon', 'chair', 'cour', 'cuiller', 'mer', 'tour', 'brebis', 'fois', 'oasis', 'souris', 'vis', 'burlat', 'dent', 'dot', 'forêt', 'mort', 'nuit', 'part', 'plupart', 'ziggourat', 'roseval', 'faim', 'croix', 'noix', 'paix', 'toux', 'voix'], matches: ['b', 'c', 'd', 'f', 'l', 'm', 'n', 'p', 'r', 's', 't', 'x'], explanation: '' },
-]
-
-export const conceptualPatterns = [
-  { key: 74, gender: 0, rule: 'Words ending in other consonants (in the spelling).', matches: ['nez', 'étang', 'sang', 'rang'], explanation: '' },
-  { key: 75, gender: 0, rule: 'Certain nouns referring to animals that can refer to only the male of the species. For example: étalon (stallion), cerf (stag), matou (tomcat).', matches: ['étalon', 'stag', 'matou'], explanation: '' },
-  { key: 76, gender: 0, rule: 'Masculine nouns that are \'generic\' terms and can refer to either a male or female of the species. For example, le cheval can refer to either a male or female horse.', matches: ['cheval'], explanation: '' },
-  { key: 77, gender: 0, rule: 'Names of towns. Other place names (departments, rivers, countries) not ending in -e. Sometimes town names, especially if they look or sound feminine (e.g. Marseilles ending in -es), can be treated as feminine. This is quite rare, though.', matches: ['Mexique', 'Combodge', 'Rhône', 'Finistère', 'Zimbabwe', 'Norvège'], explanation: '' },
-  { key: 78, gender: 0, rule: 'Nouns ending in eur, generally derived from a verb, denoting people, professions, machines carrying out an activity or scientific things: aspirateur, facteur, ordinateur', matches: ['aspirateur', 'facteur', 'ordinateur'], explanation: '' },
-  { key: 79, gender: 0, rule: 'Compound nouns of the form verb-noun: porte-monnaie, pare-brise, tire-bouchon.', matches: ['porte-monnaie', 'pare-brise', 'tire-bouchon'], explanation: '' },
-  { key: 80, gender: 0, rule: 'Nouns referring to male people.', matches: ['adulte', 'homme', 'père', 'frère', 'oncle', 'cousin', 'garçon', 'moine'], explanation: '' },
-  { key: 81, gender: 0, rule: 'A handful of nouns are masculine, whatever the gender of the person they refer to, e.g.: amateur, artiste, auteur, nudiste, témoin, vainqueur, voyou plus certain job titles.', matches: ['amateur', 'artiste', 'auteur', 'adolescent', 'nudiste', 'témoin', 'vainqueur', 'voyou'], explanation: '' },
-  { key: 82, gender: 1, rule: 'Place names ending in -e.', matches: ['Franche-Comté', 'Belize', 'Cambodge', 'Mexique', 'Mozambique', 'Zaïre', 'Zimbabwe'], explanation: '' },
-  { key: 83, gender: 1, rule: 'Nouns referring to female people.', matches: ['femme', 'fille', 'tante', 'mère', 'sœur', 'fille'], explanation: '' },
-  { key: 84, gender: 1, rule: 'These are feminine, whatever the gender of the person: personne, victime, recrue (recruit), connaissance (acquaintance).', matches: ['personne', 'victime', 'recrue', 'connaissance'], explanation: '' },
-  { key: 84, gender: 1, rule: 'These are masculine, whatever the gender of the person: témoin, guide', matches: ['témoin', 'guide', 'pygmée'], explanation: '' },
-  { key: 85, gender: 1, rule: 'Certain nouns referring to animals that can refer to only the female of the species. For example: chatte (female cat), chienne (bitch), louve (she-wolf).', matches: ['chatte', 'chienne', 'louve'], explanation: '' },
-  { key: 86, gender: 1, rule: 'Feminine nouns that are \'generic\' terms and can refer to either male or female of the species. For example, la souris can refer to either a male or female mouse.', matches: [], explanation: '' },
-  { key: 87, gender: 1, rule: 'Most other endings consisting of Vowel + Consonant + e: -ine, -ise, -alle, -elle, -esse, -ette etc', matches: ['episode', 'espace', 'intervalle', 'mille', 'magazine', 'réverbère', 'squelette'], explanation: '' },
-  { key: 88, gender: 1, rule: 'Figurative nouns ending in -eur, usually derived from an adjective: rougeur, largeur, pâleur, couleur, horreur, rumeur', matches: [], explanation: '' },
-  { key: 89, gender: 1, rule: 'Words for durations of time are feminine', matches: ['journée', 'année', 'soirée', 'matinée', 'nuitée', 'quarantaine'], explanation: '' },
-  { key: 90, gender: 1, rule: 'Words for points in time are masculine', matches: ['jour', 'an', 'soir', 'matin', 'lundi', 'mardi', 'mercredi', 'jeudi', 'vendredi', 'samedi', 'dimanche', 'les printemps', 'l\'été', 'l\'automne', 'l\'hiver', 'février', 'janvier', 'mars', 'avril', 'mai', 'juin', 'juillet', 'août', 'septembre', 'octobre', 'novembre', 'décembre'], explanation: '' },
-  { key: 91, gender: 1, rule: 'Apocopes are words which drop part of the end. Think of \'cred\' for credit or \'mag\' for magazine in English. Apocopes in French usually take the gender of their source words.', matches: ['santiag', 'métro', 'vélo', 'auto', 'accu', 'actu', 'ado', 'appart', 'ciné', 'fac', 'manif', 'météo', 'moto'], explanation: '' },
-  { key: 92, gender: 0, rule: 'Most untraslated foreign words are masculine', matches: ['look', 'jury', 'stock', 'gang', 'crash', 'gaz', 'match', 'whisky', 'URL', 'trend', 'week-end'], explanation: '' },
-  { key: 92, gender: 0, rule: 'The names of wines and most cheeses are usually masculine', matches: ['Bordeaux', 'Bourgorgne', 'Chablis', 'Pinot', 'Brie', 'Cantal', 'Camembert'], exceptions: ['tomme', 'brique'], explanation: '' },
-  { key: 92, gender: 0, rule: 'Colors are masculine', matches: [], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Numbers are masculine if they are exact, but numbers that are approximate are feminine', matches: [], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Numbers that are approximate are feminine, but numbers are masculine if they are exact', matches: ['cinquaiantaine', 'douzaine'], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Metals are masculine', matches: ['or', 'argent', 'plomb'], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Lanugages are masculine', matches: ['chinois', 'anglais', 'français', 'espagnol'], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Trees are masculine, and so are most shrubs but vines are feminine.', matches: ['arbre', 'pommier', 'pin', 'sapin', 'saule', 'platane'], exceptions: ['vigne'], explanation: '' },
-  { key: 92, gender: 1, rule: 'Most brand names of cars and watches', matches: ['Rolex', 'Swatch', 'Peugeot', 'Citroën', 'Mercedes', 'BMW'], exceptions: [], explanation: '' },
-  { key: 92, gender: 0, rule: 'Units of measure', matches: ['kilo', 'mètre', 'joule', 'hectare', 'litre', 'quart'], exceptions: ['moitié'], explanation: '' },
-  { key: 92, gender: 1, rule: 'The proper names of most rivers', matches: [], exceptions: [], explanation: '' },
-  { key: 92, gender: 1, rule: 'The proper names of domains of learning', matches: ['science', 'géographie', 'astronomie', 'histoire', 'chimie'], exceptions: ['droit'], explanation: '' },
-];
